requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
sqlmodel>=0.0.8
asyncpg>=0.27
aiohttp>=3.8
python-dotenv>=0.21
pydantic>=1.10
pyjwt>=2.8
qrcode>=7.3
pillow>=9.5
numpy>=1.25
pandas>=2.2
scikit-learn>=1.2
statsmodels>=0.14
tensorflow>=2.13; python_version >= "3.9"  # optional LSTM
prophet>=1.1; python_version >= "3.9"  # optional Prophet (fbprophet)
python-multipart>=0.0.6
cryptography>=40.0